Fix pathological backtracking in signature regexp
This commit fixes the pathological backtracking in the regexp so that we
don't need the RE2 gem anymore.

The problem was the `(\w+\s*){1,3}` part.  Since we accept 0 or more
spaces, and 1 to 3 times, the regex engine would keep backtracking to
figure out if it should do 1, 2, or 3 matches.  We know we want at
least one match, and it *must* be followed by one or more spaces.  If we
remove the space before the "y", we can eliminate the backtracking by
making at least one space required (as the original regular expression
indicated)