handle mulitline reply headers in Gmail
Gmail will break the reply header into multiple lines when it is
over 80 characters. As an example:

On Fri, Apr 3, 2020 at 12:32 PM Someone <someone@example.com> wrote:

Will become:

On Fri, Apr 3, 2020 at 12:32 PM Someone <
someone@example.com> wrote:

This was causing the reply header to be included in the final
returned body. This update matches the reply header across
multiple lines so it is excluded.